deduplicate test helpers into tests/utils.py
