+++ tests/test_bound_interface.py
@@ -1,13 +1,11 @@
 import numpy as np
 import pytest
 from numpy import linalg
-import numpy.testing as npt
 import itertools
-from utils import get_rstate, get_printing
+from utils import get_rstate, get_printing, check_results_gau
 
 import dynesty  # noqa
 import dynesty.bounding as db
-from dynesty import utils as dyfunc  # noqa
 """
 Run a series of basic tests to check whether anything huge is broken.
 
@@ -51,51 +49,6 @@ class Box(db.Bound):
         self.logvol = np.log(self.size) * self.ndim
 
 
-def bootstrap_tol(results, rstate):
-    """ Compute the uncertainty of means/covs by doing bootstrapping """
-    n = len(results['logz'])
-    niter = 50
-    pos = results.samples
-    wts = results.importance_weights()
-    means = []
-    covs = []
-
-    for i in range(niter):
-        sub = rstate.uniform(size=n) < wts / wts.max()
-        ind0 = np.nonzero(sub)[0]
-        ind1 = rstate.choice(ind0, size=len(ind0), replace=True)
-        mean = pos[ind1].mean(axis=0)
-        cov = np.cov(pos[ind1].T)
-        means.append(mean)
-        covs.append(cov)
-    return np.std(means, axis=0), np.std(covs, axis=0)
-
-
-def check_results(results,
-                  mean_truth,
-                  cov_truth,
-                  logz_truth,
-                  mean_tol,
-                  cov_tol,
-                  logz_tol,
-                  sig=4):
-    """ Check if means and covariances match match expectations
-    within the tolerances
-
-    """
-    results.summary()
-    pos = results.samples
-    wts = np.exp(results['logwt'] - results['logz'][-1])
-    assert np.allclose(results.importance_weights(), wts)
-    mean, cov = dyfunc.mean_and_cov(pos, wts)
-    logz = results['logz'][-1]
-    logzerr = results['logzerr'][-1]
-    assert logzerr < 10  # check that it is not too large
-    npt.assert_array_less(np.abs(mean - mean_truth), sig * mean_tol)
-    npt.assert_array_less(np.abs(cov - cov_truth), sig * cov_tol)
-    npt.assert_array_less(np.abs((logz_truth - logz)), sig * logz_tol)
-
-
 # GAUSSIAN TEST
 
 
@@ -135,24 +88,6 @@ class Gaussian:
         u[:] = -np.ones(len(u))
 
         return ret
-
-
-def check_results_gau(results, g, rstate, sig=4, logz_tol=None):
-    if logz_tol is None:
-        logz_tol = sig * results['logzerr'][-1]
-    mean_tol, cov_tol = bootstrap_tol(results, rstate)
-    # just check that resample_equal works
-    dyfunc.resample_equal(results.samples,
-                          np.exp(results['logwt'] - results['logz'][-1]))
-    results.samples_equal()
-    check_results(results,
-                  g.mean,
-                  g.cov,
-                  g.logz_truth,
-                  mean_tol,
-                  cov_tol,
-                  logz_tol,
-                  sig=sig)
 
 
 # try all combinations except
