+++ tests/test_sampler_interface.py
@@ -1,10 +1,9 @@
 import numpy as np
 import pytest
 from numpy import linalg
-import numpy.testing as npt
 import itertools
 
-from utils import get_rstate, get_printing
+from utils import (get_rstate, get_printing, bootstrap_tol, check_results)
 import dynesty.internal_samplers as dysa
 import dynesty  # noqa
 from dynesty import utils as dyfunc  # noqa
@@ -15,51 +14,6 @@ Run a series of basic tests to check whether anything huge is broken.
 
 nlive = 500
 printing = get_printing()
-
-
-def bootstrap_tol(results, rstate):
-    """ Compute the uncertainty of means/covs by doing bootstrapping """
-    n = len(results['logz'])
-    niter = 50
-    pos = results.samples
-    wts = results.importance_weights()
-    means = []
-    covs = []
-
-    for i in range(niter):
-        sub = rstate.uniform(size=n) < wts / wts.max()
-        ind0 = np.nonzero(sub)[0]
-        ind1 = rstate.choice(ind0, size=len(ind0), replace=True)
-        mean = pos[ind1].mean(axis=0)
-        cov = np.cov(pos[ind1].T)
-        means.append(mean)
-        covs.append(cov)
-    return np.std(means, axis=0), np.std(covs, axis=0)
-
-
-def check_results(results,
-                  mean_truth,
-                  cov_truth,
-                  logz_truth,
-                  mean_tol,
-                  cov_tol,
-                  logz_tol,
-                  sig=4):
-    """ Check if means and covariances match match expectations
-    within the tolerances
-
-    """
-    results.summary()
-    pos = results.samples
-    wts = np.exp(results['logwt'] - results['logz'][-1])
-    assert np.allclose(results.importance_weights(), wts)
-    mean, cov = dyfunc.mean_and_cov(pos, wts)
-    logz = results['logz'][-1]
-    logzerr = results['logzerr'][-1]
-    assert logzerr < 10  # check that it is not too large
-    npt.assert_array_less(np.abs(mean - mean_truth), sig * mean_tol)
-    npt.assert_array_less(np.abs(cov - cov_truth), sig * cov_tol)
-    npt.assert_array_less(np.abs((logz_truth - logz)), sig * logz_tol)
 
 
 class TestSampler(dysa.InternalSampler):
